requirements.txt
tensorflow-gpu
numpy
opencv-python-headless
matplotlib
glob2
scipy
